Fix: StartTimes & Offsets -> Floats
Former-commit-id: 0df596e2dea61062d6411c8772090ba36c26bae5 [formerly 10dd85a77c636e8409ca407a943e8cd267477565]
Former-commit-id: 079e3771309f580f7a4fd0f0697d88a4cc9906e3

+++ Assets/Scripts/Osu.Beatmap/Structures/TimingPoint.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Osu.Beatmap {
 
     public struct TimingPoint {
-        public int Offset;
+        public float Offset;
         public float MillisecondsPerBeat;
         public int Meter;
         public int SampleType;
